// src/types.ts
export interface SCPEntry {
  id: string;
  scp_designation: string;
  code_name: string;
  description: string;
  ontological_category: string;
  morphology: string;
  object_containment_class: string;
  disruption_class: string;
  risk_class: string;
  rating: number;
  sapience_level: string;
  security_clearance_level: string;
  emotional_tone: string;
  year_discovered: string;
  date_discovered: string;
  country_of_discovery: string;
  city_location_details: string;
  discovery_details: string;
  discovery_method: string;
  origin_type: string;
  origin_detail: string;
  historical_real_world_associations: string;
  original_scp_article_author: string;
  special_containment_protocol: string;
  containment_facility: string;
  containment_facility_type: string;
  containment_status: string;
  challenges_to_containment_type: string;
  challenges_to_containment_detail: string;
  termination_attempts: string;
  trigger_detail: string;
  normal_behaviour_detail: string;
  normal_behaviour_type: string;
  triggered_behaviour_detail: string;
  triggered_behaviour_type: string;
  respective_department: string;
  involved_doctors: string;
  involved_researchers: string;
  involved_agents: string;
  involved_directors: string;
  other_involved_staff: string;
  involved_mobile_task_force: string;
  involved_mobile_task_force_codename: string;
  related_groups: string;
  o5_or_administrator_involvement: string;
  anomaly_type: string;
  affected_natural_laws: string;
  anomalous_effects: string;
  anomalous_effects_types: string;
  anomalous_secondary_manifestations: string;
  interaction_type: string;
  ability_types: string;
  ability_details: string;
  spread_mechanism: string;
  weaknesses: string;
  resistances: string;
  physical_attributes: string;
  psychological_traits: string;
  speed_of_pursuit: string;
  time_till_effect: string;
  range_of_effects: string;
  fatality_rate: string;
  casualty_records: string;
  related_k_class_scenarios: string;
  related_scps: string;
  locations_of_interest: string;
  locations_of_interest_types: string;
  persons_of_interest: string;
  groups_of_interest: string;
  experiment_logs: string;
  exploration_logs: string;
  incident_reports: string;
  interviews: string;
  witness_disposition_records: string;
  associated_artifacts: string;
  associated_documents: string;
  total_number_of_anomalous_entities: string;
  anomalies_spawned: string;
  source_link: string;
  scp_image_url: string | null;
  scp_image2_url: string | null;
  image_source: string;
  image2_source: string;
  scp_audio_url: string | null;
  audio_source: string;
  official_cover_story: string;
  tag_list: string;
  image_urls?: string[] | null;
  image_sources?: string[] | null;
  audio_urls?: string[] | null;
  audio_sources?: string[] | null;
  video_urls?: string[] | null;
  video_sources?: string[] | null;
  created_at: string;
  updated_at: string;
}